Option to trim images (#12)
* New Configuration Option for Trimming Images
Added a new config option to trim images, allowing users to customize their image display.

* Image Model Update and URL Query String Addition
Updated the image model with the new configuration value, which is now included as an extra parameter in the URL query string when requesting Imgix for image processing. This makes it easier for users to apply customized image trimming while processing.